JumpingOnTheClouds problem completed. Used easily understood variables "movesMade, curentPlace, placePlusOne, placePlusTwo, endReached" to keep a record of every process used. Method uses a While loop that determins if endReached is still false, it continues.
Conditional statements used  to determin best move to make.
If a jump of 2 can be made, it will be, otherwise 1. If neither are appropiate, it means the end was reached and endReached is flipped to "true", ending the loop.
Every time a move was made, it was counted. At the end, the total moves made was returned.